+++ internal/provider/resource_workspace_settings_test.go
@@ -16,6 +16,7 @@ func TestAccWorkspaceSettingsResourceDefault(t *testing.T) {
 				Config: testAccWorkspaceSettingsResourceConfigDefault(),
 				Check: resource.ComposeAggregateTestCheckFunc(
 					resource.TestMatchResourceAttr("linear_workspace_settings.test", "id", uuidRegex()),
+					resource.TestCheckResourceAttr("linear_workspace_settings.test", "allow_members_to_invite", "true"),
 					resource.TestCheckResourceAttr("linear_workspace_settings.test", "enable_roadmap", "false"),
 					resource.TestCheckResourceAttr("linear_workspace_settings.test", "enable_git_linkback_messages", "true"),
 					resource.TestCheckResourceAttr("linear_workspace_settings.test", "enable_git_linkback_messages_public", "false"),
@@ -32,6 +33,7 @@ func TestAccWorkspaceSettingsResourceDefault(t *testing.T) {
 				Config: testAccWorkspaceSettingsResourceConfigDefault(),
 				Check: resource.ComposeAggregateTestCheckFunc(
 					resource.TestMatchResourceAttr("linear_workspace_settings.test", "id", uuidRegex()),
+					resource.TestCheckResourceAttr("linear_workspace_settings.test", "allow_members_to_invite", "true"),
 					resource.TestCheckResourceAttr("linear_workspace_settings.test", "enable_roadmap", "false"),
 					resource.TestCheckResourceAttr("linear_workspace_settings.test", "enable_git_linkback_messages", "true"),
 					resource.TestCheckResourceAttr("linear_workspace_settings.test", "enable_git_linkback_messages_public", "false"),
@@ -42,6 +44,7 @@ func TestAccWorkspaceSettingsResourceDefault(t *testing.T) {
 				Config: testAccWorkspaceSettingsResourceConfigNonDefault(),
 				Check: resource.ComposeAggregateTestCheckFunc(
 					resource.TestMatchResourceAttr("linear_workspace_settings.test", "id", uuidRegex()),
+					resource.TestCheckResourceAttr("linear_workspace_settings.test", "allow_members_to_invite", "false"),
 					resource.TestCheckResourceAttr("linear_workspace_settings.test", "enable_roadmap", "true"),
 					resource.TestCheckResourceAttr("linear_workspace_settings.test", "enable_git_linkback_messages", "false"),
 					resource.TestCheckResourceAttr("linear_workspace_settings.test", "enable_git_linkback_messages_public", "true"),
@@ -68,6 +71,7 @@ func TestAccWorkspaceSettingsResourceNonDefault(t *testing.T) {
 				Config: testAccWorkspaceSettingsResourceConfigNonDefault(),
 				Check: resource.ComposeAggregateTestCheckFunc(
 					resource.TestMatchResourceAttr("linear_workspace_settings.test", "id", uuidRegex()),
+					resource.TestCheckResourceAttr("linear_workspace_settings.test", "allow_members_to_invite", "false"),
 					resource.TestCheckResourceAttr("linear_workspace_settings.test", "enable_roadmap", "true"),
 					resource.TestCheckResourceAttr("linear_workspace_settings.test", "enable_git_linkback_messages", "false"),
 					resource.TestCheckResourceAttr("linear_workspace_settings.test", "enable_git_linkback_messages_public", "true"),
@@ -84,6 +88,7 @@ func TestAccWorkspaceSettingsResourceNonDefault(t *testing.T) {
 				Config: testAccWorkspaceSettingsResourceConfigNonDefault(),
 				Check: resource.ComposeAggregateTestCheckFunc(
 					resource.TestMatchResourceAttr("linear_workspace_settings.test", "id", uuidRegex()),
+					resource.TestCheckResourceAttr("linear_workspace_settings.test", "allow_members_to_invite", "false"),
 					resource.TestCheckResourceAttr("linear_workspace_settings.test", "enable_roadmap", "true"),
 					resource.TestCheckResourceAttr("linear_workspace_settings.test", "enable_git_linkback_messages", "false"),
 					resource.TestCheckResourceAttr("linear_workspace_settings.test", "enable_git_linkback_messages_public", "true"),
@@ -94,6 +99,7 @@ func TestAccWorkspaceSettingsResourceNonDefault(t *testing.T) {
 				Config: testAccWorkspaceSettingsResourceConfigDefault(),
 				Check: resource.ComposeAggregateTestCheckFunc(
 					resource.TestMatchResourceAttr("linear_workspace_settings.test", "id", uuidRegex()),
+					resource.TestCheckResourceAttr("linear_workspace_settings.test", "allow_members_to_invite", "true"),
 					resource.TestCheckResourceAttr("linear_workspace_settings.test", "enable_roadmap", "false"),
 					resource.TestCheckResourceAttr("linear_workspace_settings.test", "enable_git_linkback_messages", "true"),
 					resource.TestCheckResourceAttr("linear_workspace_settings.test", "enable_git_linkback_messages_public", "false"),
@@ -120,6 +126,7 @@ resource "linear_workspace_settings" "test" {
 func testAccWorkspaceSettingsResourceConfigNonDefault() string {
 	return `
 resource "linear_workspace_settings" "test" {
+	allow_members_to_invite = false
 	enable_roadmap = true
 	enable_git_linkback_messages = false
 	enable_git_linkback_messages_public = true